updated arm64 in every target

+++ CU_iOS/CampusPublishers/UserInterface/Classes/Controllers/Page/CPGalleryViewController.h
@@ -29,7 +29,7 @@
     CPSwipeView *swipeView;
     CPFooterItemType footerType;
     CPDataManger *dtManager;
-    int currentIndex;
+    NSInteger currentIndex;
     UIImage *imgPlaceholder;
     NSOperationQueue *loadImagesOperationQueue;
     AFOpenFlowView *openFlow;
